--- accounts/urls.py
@@ -1,9 +1,0 @@
-from django.urls import path
-
-from . import views
-
-urlpatterns = [
-  path('register', views.register, name='register'),
-  path('login/', views.login, name='login'),
-  path('logout/', views.logout, name='logout'),
-]
