+++ beemer_boys/settings.py
@@ -39,7 +39,6 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'motorcycle',
     'stripe',
-    'accounts',
 ]
 
 MIDDLEWARE = [
