+++ beemer_boys/urls.py
@@ -16,12 +16,10 @@ Including another URLconf
 from django.contrib import admin
 from django.urls import path, include
 from motorcycle import views
+from django.conf import settings
+from django.conf.urls.static import static
 
 urlpatterns = [
   path('admin/', admin.site.urls),
   path('', include ('motorcycle.urls')),
-  path('accounts/', include('accounts.urls')),
-  
- 
-  
-]
+] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
