+++ motorcycle/templates/base.html
@@ -20,16 +20,9 @@
   <header>
     <span class="bb-logo">BEEMER BOYS</span>
     <ul class='header'>
-      {% if user.is_authenticated %}
-        <li><a class="nav-link" href="{% url 'about' %}">ABOUT</a></li>
-        <li><a class="nav-link" href="{% url 'models' %}">MODELS</a></li>
-        <li><a class="nav-link" href="{% url 'logout' %}">LOGOUT</a></li>
-      {% else %}
-        <li><a class="nav-link" href="{% url 'about' %}">ABOUT</a></li>
-        <li><a class="nav-link" href="{% url 'models' %}">MODELS</a></li>
-        <li><a class="nav-link" href="{% url 'login' %}">SIGN IN</a></li>
-        <li><a class="nav-link" href="{% url 'register' %}">REGISTER</a></li>
-      {% endif %}
+      <li><a class="nav-link" href="{% url 'about' %}">ABOUT</a></li>
+      <li><a class="nav-link" href="{% url 'models' %}">MODELS</a></li>
+      <li><a class="nav-link" href="{% url 'login' %}">SIGN IN</a></li>
     </ul>
   </header>
 {% block content %}
