+++ motorcycle/templates/form.html
@@ -15,7 +15,7 @@
         <input type="password" name="password" placeholder="Password" minlength="4">
         <input type="submit" name="" value="login">
       </form>
-      <a href="#"> Forget Password</a> <a href="{% url 'register' %}">Register</a>
+      <a href="#"> Forget Password</a>
     </div>
   </div>
 </section>
