+++ motorcycle/templates/modelinfo.html
@@ -17,6 +17,5 @@
           data-locale="auto">
   </script>
 </form>
-
 {% endblock %}
